Refactor users' controller to own package
