+++ internal/router/router.go
@@ -7,6 +7,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/cors"
 	"github.com/source-academy/stories-backend/controller"
+	"github.com/source-academy/stories-backend/controller/users"
 	"github.com/source-academy/stories-backend/internal/config"
 	"github.com/source-academy/stories-backend/internal/utils/constants"
 )
@@ -38,9 +39,9 @@ func Setup(config *config.Config, injectMiddleWares []func(http.Handler) http.Ha
 	})
 
 	r.Route("/users", func(r chi.Router) {
-		r.Get("/", controller.GetUsers)
-		r.Get("/{userID}", controller.GetUser)
-		r.Post("/", controller.CreateUser)
+		r.Get("/", users.GetUsers)
+		r.Get("/{userID}", users.GetUser)
+		r.Post("/", users.CreateUser)
 	})
 
 	return r
